XWIKI-6665: Saving edits after lock expiration doesn't warn the user that the page may be stale
  * SaveAction now looks the content diff to avoid the attachment issue.
A conflict is detected only if the diff changed between the current
document and latest version. The diff is computed both on the content
and on the objects since we want to support inline edition.
  * Add some methods in xwiki-meta.js to bump and set a version. This
fires an event "xwiki:document:changeVersion"
  * dataeditors.js now bump the document version when an object is added
or removed to avoid false positive for the conflict detection
  * actionButton.js now listen to the changeVersion event to update the
current version of the doc.